+++ app/src/components/Tasks/index.tsx
@@ -1,10 +1,6 @@
 import { Flex } from '@chakra-ui/react'
 import type { Task } from '../../models/task'
-import dayjs from 'dayjs'
-import duration from 'dayjs/plugin/duration'
 import { TaskContainer } from './TaskContainer'
-
-dayjs.extend(duration)
 
 export const Tasks = ({ tasks }: { tasks: Task[] }) => {
   const displayWidth = '600px'
